requirements.txt
appnope==0.1.0
attrs==19.3.0
backcall==0.1.0
behave==1.2.6
behave-pandas==0.3.0
click==7.0.0
decorator==4.4.1
entrypoints==0.3
gherkin-official==4.1.3
importlib-metadata==1.4.0
jedi==0.15.2
more-itertools==8.1.0
numpy==1.18.1
packaging==20.0
pandas==0.25.3
parse==1.14.0
parse-type==0.5.2
parso==0.5.2
pexpect==4.7.0
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==2.0.10
ptyprocess==0.6.0
py==1.8.1
pyflakes==2.1.1
Pygments==2.5.2
pyparsing==2.4.6
pytest==5.3.3
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
tabulate==0.8.6
traitlets==4.3.3
wcwidth==0.1.8
zipp==1.0.0
